+++ src/main/java/com/example/doantotnghiepbe/dto/ReportDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class ReportDTO {
     private Integer reportId;
-    private Integer user;
+    private Long user;
     private Integer post;
     private String reportType;
     private String reportContent;
