+++ src/main/java/com/example/doantotnghiepbe/repository/PaymentRepository.java
@@ -14,7 +14,7 @@ public interface PaymentRepository extends JpaRepository<Payment, Long> {
     Payment findByVnpTxnRef(String vnpTxnRef);
 
     @Query("SELECT p FROM Payment p WHERE p.postId.user.id = :userId")
-    List<Payment> findPaymentsByUserId(@Param("userId") Integer userId);
+    List<Payment> findPaymentsByUserId(@Param("userId") Long userId);
 
 
 }
